Reworked Minge Weapons, added SWEPs, prepared HUD files, and renamed minge_base to base_minge
I think a base prefix for base entities makes it easier to differentiate between what is intended for use and what is not. I am now also working on the required SWEPs for this gamemode. In the end there should be 4 weapons the player has: primary weapon, secondary weapon, wrench, and Minge Defense PDA