Internals for wave logic started, more hammer entities, and file structure for map configs/scripts
Also added a font file, but I have no way of testing if it works right now. I added the font (Franklin Gothic Heavy) because my battle buddy didn't have the font already installed on his computer (even though it comes with Windows 10 :/ ). I plan to make as much as possible done inside of hammer so this doesn't turn in to nZombies like map configuration. If most of the configuration is done in hammer we don't have to worry about linging stuff up, clipping it, etc. we can just worry about things like waves and stuff. The only thing I really want outside of hammer is the configuration of the waves (mainly enemies in waves), the configuration of path targets' order (not their positions, we are doing this outside to make it easier on my brain), and active spawn points in use.

Anyways, the file structure will accommodate custom scripts, but what will mainly be used is configurations generated with an in-game tool. The custom scripts will allow you to override certain wave's generation so you can balance even further, or add interactive events.

Also I'm still trying to figure out why ui/status isn't loading correctly... probably has to do with the init crap.