Prep for localization, mdinfo_screen improvements and testing, + more
Localization will be done through the built in system + a string formatter with keys like "mingedefense.welcome=Welcome [player] to Minge Defense!" is written in the localization file but gets fromatted in game into "Welcome Cryotheum to Minge Defense!"

I'm also making an icon generator for the minges for the wave screen, so I don't have to manually make textures. This should also save on file size.

I made a different player class to experiment with rendering 3D/3D2D stuff on players, and I found that the best way is to render an entity parented on the player.