Team status bar and rewrote wave code
I also changed the hooks around a bit; the wiki page will be updated when I have time to do so. With the partial rewrite of the wave code, the folder was renamed to round. This is because when I was setting up code for transitioning between waves and difficulty it became awkward to use the Wave namespace when programming stuff outside of a wave. So with the namespace for Round, we can make the code a bit easier for me and probably others to understand. It also allows the wave related files to focuse more on the wave itself than the round, meaning (hopefully) less headaches down the line.

Language also had a bit of and overhaul. There are fields that can get parsed into the desired text now, so languages like french are not locked down to the normal string.format.

I've been also thinking about using the library functions as methods instead of indexed functions. This would make the code cleaner, but my IDE does not like it as much :(.